aarch64/gic: check before accessing group0 interrupt related registers
On a system with two security states, We have to check whether group0
interrupts belong to the Secure state before accessing its related
registers. Otherwise, it can result in stall.

Check for whether EL3 exists first. If it exists, and GICD_CTLR.DS is not
set, it means that group0 interrupts belong to the Security state. We
should not touch its related registers. Otherwise, we can access them.