Kconfig: switch to Kconfig for defining CONFIG_ options
We are to support multiple architectures. However, the current CONFIG_*
options management is too limited. We are likely to need more that yes/no
options in the future. In addition, we need a better visualization of
CONFIG_* option dependencies.

This patch introduces the use of Kconfig for CONFIG_ option management.
Kconfig is parsed, and its corresponding .config file is generated by
kconfiglib python scripts. With kconfiglib, we can have a modern
menuconfig similar to the one we see in Linux kernel project. We can also
add descriptive comments on each option in the future as well.

Note that the project will depend on python and ncurses (for menuconfig)
starting from this patch.