drivers/pci: add bus number range check function
This commit adds a range check function to determine if the bus number
is within the bounds of the MCFG. In addition, the PCI driver falls
back to the port-mapped I/O access if the function returns false. With
this change, "config_mmio" field of "pci_device" struct is no longer
necessary.