xhci: support USB_HOOK_DISCARD
"usb_hook_process()" returns an "int" value (e.g., USB_HOOK_DISCARD,
which is 0x20000), but the xHCI driver previously stored the returned
value in a "u8" variable. As a result, the driver could not handle the
variable as expected.

This commit changes the driver to store the returned value in an "int"
variable. With this commit, the driver can handle the variable
returned from "usb_hook_process()" correctly.