aarch64/cnt: add functions to access system counter and frequency
To read the system counter properly, we need isb instruction before actual
the reading with mrs instruction to avoid out-of-order execution. It is a
good idea to add cnt_get_cntpct_el0() as a convinient function. It hides
away the need for the isb instruction.

For styling consistency, this patch also adds cnt_get_cntfrq_el0() for
reading the clock frequency as well.