svm: use nRIP for vmmcall, cpuid, clgi, stgi, vmrun and invlpga instructions
Previously, these instructions were not emulated correctly when they
had one or more prefixes, due to the use of fixed offset addition.
For example, instructions such as 'rep; cpuid' in assembly language
triggered the issue.  This commit changes the implementation to use
nRIP to fix the issue.  If nRIP is not supported, the previous
addition method is used as a fallback.  As recent processors support
nRIP, the issue will only affect older processors.