usb: clean up new device handling
The UHCI and EHCI drivers use a hook to find SetAddress.  The xHCI
driver sets an internal flag in a slot meta structure after the xHCI's
Address Device Command and invokes usb_device to handle new device
when the first doorbell is rung for that slot.  A control message
following SetAddress can now properly be hooked in xHCI too.