ehci: improve doorbell handling
The Interrupt on Async Advance Doorbell bit (bit 6) in the USB Command
Register should not be set if the Asynchronous Schedule Enable bit
(bit 5) is not set.  Furthermore, if the Run/Stop bit (bit 0) is not
set, setting bit 6 is unnecessary.

This changeset ensures bit 6 is set only when both bit 5 and bit 0 are
set.