virtio_net: check command size in MAC address filtering
The current `process_ctrl_mac_cmd()` code checks the size of `cmd` >=
11. This value is when both unicast and multicast address entries is 0.
However, it does not check the `cmd` has enough size for each address
entries >= 1. This causes out-of-bounds read if a crafted command sent
where `cmd_size` is 11 and set each entries >= 1. The read result will
be written to `vnet`, but there is no way to read the MAC address
filtering values from the guest driver, the guest cannot read the leaked
memory contents.

Signed-off-by: Akira Moroo <retrage01@gmail.com>